chore: updates split secret location, also uses key based approach for identifier/keypackage mapping
